starter_agent_scaffold/requirements.txt
# Core
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
PyYAML
httpx

# LLM Providers (install only the one you need)
# openai
# google-generativeai
# anthropic

# Integrations (install only the ones you need)
# slack_bolt

# Memory
# redis

# Testing
pytest
pytest-mock
pytest-asyncio

# Security & Quality
flake8
bandit

# Observability
prometheus-client
